requirements.txt
amqp==5.0.2
anyjson==0.3.3
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.5
django-filter==2.4.0
django-restql==0.11.0
djangorestframework==3.12.2
generics==3.2.1
idna==2.10
kombu==5.0.2
Markdown==3.3.3
prompt-toolkit==3.0.10
pyPEG2==2.15.2
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
